refactor: split database.py into modular package structure
