+++ app/database/config.py
@@ -1,0 +1,9 @@
+import mariadb
+
+db_config = {
+    "host": "myapidb",
+    "port": 3306,
+    "user": "myapi",
+    "password": "myapi",
+    "database": "myapi",
+}
